+++ cf/test/test_external.py
@@ -1,6 +1,5 @@
 import datetime
 import faulthandler
-import inspect
 import os
 import tempfile
 import unittest
@@ -46,9 +45,6 @@ class ExternalVariableTest(unittest.TestCase):
         os.remove(self.tempfilename_external)
 
     def test_EXTERNAL_READ(self):
-        if self.test_only and inspect.stack()[0][3] not in self.test_only:
-            return
-
         # Read the parent file on its own, without the external file
         f = cf.read(self.parent_file, verbose=0)
 
@@ -61,7 +57,7 @@ class ExternalVariableTest(unittest.TestCase):
 
         f = f[0]
 
-        cell_measure = f.constructs.filter_by_identity("measure:area").value()
+        cell_measure = f.cell_measure("measure:area")
 
         self.assertTrue(cell_measure.nc_get_external())
         self.assertEqual(cell_measure.nc_get_variable(), "areacella")
@@ -78,9 +74,7 @@ class ExternalVariableTest(unittest.TestCase):
             str(i)
             i.dump(display=False)
 
-        cell_measure = (
-            f[0].constructs.filter_by_identity("measure:area").value()
-        )
+        cell_measure = f[0].cell_measure("measure:area")
 
         self.assertEqual(len(f), 1)
         self.assertEqual(len(c), 1)
@@ -121,9 +115,6 @@ class ExternalVariableTest(unittest.TestCase):
             self.assertTrue(c[i].equals(f[i], verbose=2))
 
     def test_EXTERNAL_WRITE(self):
-        if self.test_only and inspect.stack()[0][3] not in self.test_only:
-            return
-
         parent = cf.read(self.parent_file)
         combined = cf.read(self.combined_file)
 
@@ -138,7 +129,7 @@ class ExternalVariableTest(unittest.TestCase):
         for i in range(len(g)):
             self.assertTrue(combined[i].equals(g[i], verbose=2))
 
-        cell_measure = g[0].constructs("measure:area").value()
+        cell_measure = g[0].cell_measure("measure:area")
 
         self.assertFalse(cell_measure.nc_get_external())
         cell_measure.nc_set_external(True)
@@ -146,12 +137,7 @@ class ExternalVariableTest(unittest.TestCase):
         self.assertTrue(cell_measure.properties())
         self.assertTrue(cell_measure.has_data())
 
-        self.assertTrue(
-            g[0]
-            .constructs.filter_by_identity("measure:area")
-            .value()
-            .nc_get_external()
-        )
+        self.assertTrue(g[0].cell_measure("measure:area").nc_get_external())
 
         cf.write(
             g,
@@ -176,49 +162,47 @@ class ExternalVariableTest(unittest.TestCase):
             self.assertTrue(external[i].equals(h[i], verbose=2))
 
     def test_EXTERNAL_AGGREGATE(self):
-        if self.test_only and inspect.stack()[0][3] not in self.test_only:
-            return
-
-        # Read parent file without the external file, taking first field to test
+        # Read parent file without the external file, taking first
+        # field to test
         f = cf.read(self.parent_file, verbose=0)[0]
         measure_name = "measure:area"
 
-        # Split f into parts (take longitude with 3 x 3 = 9 points) so can
-        # test the difference between the aggregated result and original f.
-        # Note all parts retain the external variable cell measure.
+        # Split f into parts (take longitude with 3 x 3 = 9 points) so
+        # can test the difference between the aggregated result and
+        # original f.  Note all parts retain the external variable
+        # cell measure.
         f_lon_thirds = [f[:, :3], f[:, 3:6], f[:, 6:]]
 
         g = cf.aggregate(f_lon_thirds)
 
         self.assertEqual(len(g), 1)
 
-        # Check cell measure construct from external variable has been removed
-        self.assertFalse(g[0].cell_measures())
+        # Check cell measure construct from external variable has been
+        # retained
+        self.assertEqual(len(g[0].cell_measures()), 1)
 
-        # Check aggregated field is identical to original with measure removed
-        f0 = f.copy()
-        f0.del_construct(measure_name)
-        self.assertEqual(g[0], f0)
+        # Check aggregated field is identical to original
+        self.assertEqual(g[0], f)
 
-        # Also check aggregation did not remove the measure from the inputs
+        # Also check aggregation did not remove the measure from the
+        # inputs
         for part in f_lon_thirds:
-            cell_measure = part.constructs.filter_by_identity(
-                "measure:area"
-            ).value()
+            cell_measure = part.cell_measure("measure:area")
             self.assertTrue(cell_measure.nc_get_external())
 
-        # Now try aggregating when one part doesn't have the cell measure,
-        # expecting all of the parts to still aggregate back to one field
-        # without the external measure (rather than  2->1 => 1 + 1 = 2 fields).
+        # Now try aggregating when one part doesn't have the cell
+        # measure
         f_lon_thirds[1].del_construct(measure_name)
         g = cf.aggregate(f_lon_thirds)
-        self.assertEqual(len(g), 1)
-        self.assertFalse(g[0].cell_measures())
+        self.assertEqual(len(g), 2)
+        self.assertFalse(g[1].cell_measures())
 
-        # Also check measure was not removed from, or added to, any input
+        # Also check measure was not removed from, or added to, any
+        # input
         for part in [f_lon_thirds[0], f_lon_thirds[2]]:
-            cm = part.constructs.filter_by_identity("measure:area").value()
+            cm = part.cell_measure("measure:area")
             self.assertTrue(cm.nc_get_external())
+
         self.assertFalse(f_lon_thirds[1].cell_measures())