+++ Changelog.rst
@@ -5,9 +5,17 @@ version 3.14.1
 
 * New keyword parameter to `cf.Data.concatenate`: ``relaxed_units``
   (https://github.com/NCAS-CMS/cf-python/issues/602)
+* New method: `cf.Field.del_properties`
+  (https://github.com/NCAS-CMS/cf-python/issues/598)
+* New keyword parameter to `cf.unique_constructs`:
+  ``ignore_properties``
+  (https://github.com/NCAS-CMS/cf-python/issues/597)
+* Allow `cf.aggregate` and to retain external cell meausures
+  (https://github.com/NCAS-CMS/cf-python/issues/605)
 * Fixed bug in `cf.Field.collapse` that always retained invalid
   vertical coordinate reference systems
   (https://github.com/NCAS-CMS/cf-python/issues/599)
+* Changed dependency: ``1.10.0.3<=cfdm<1.10.1.0``
 
 ----